modules/BaseDeDonnees/requirements.txt
Flask==1.0.2
gunicorn==19.9.0
beautifulsoup4==4.7.1
mysql-connector==2.1.6
google-cloud-storage==1.13.0
nltk==3.4
sklearn
numpy==1.16.2
